# backend/docs/swagger-output.json
{
  "openapi": "2.0",
  "info": {
    "title": "GLY VTU API",
    "description": "Auto-generated Swagger specification for GLY VTU backend.",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Use: `Bearer <token>`"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/wallet/balance": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "summary": "Get wallet balance",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Balance",
            "schema": {
              "$ref": "#/definitions/WalletBalance"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Wallet not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/wallet/send": {
      "post": {
        "tags": [
          "Wallet"
        ],
        "summary": "Send money to bank or internal user",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WalletSendRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transfer created",
            "schema": {
              "$ref": "#/definitions/WalletSendResponse"
            }
          },
          "400": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/wallet/receive": {
      "post": {
        "tags": [
          "Wallet"
        ],
        "summary": "Request money from another user",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WalletReceiveRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request created",
            "schema": {
              "$ref": "#/definitions/WalletReceiveResponse"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/balance": {
      "get": {
        "tags": [
          "Wallet"
        ],
        "summary": "Get wallet balance",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Balance",
            "schema": {
              "$ref": "#/definitions/WalletBalance"
            }
          },
          "404": {
            "description": "Wallet not found",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/send": {
      "post": {
        "tags": [
          "Wallet"
        ],
        "summary": "Send money to bank or internal user",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WalletSendRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transfer created",
            "schema": {
              "$ref": "#/definitions/WalletSendResponse"
            }
          },
          "400": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/receive": {
      "post": {
        "tags": [
          "Wallet"
        ],
        "summary": "Request money from another user",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/WalletReceiveRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request created",
            "schema": {
              "$ref": "#/definitions/WalletReceiveResponse"
            }
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "definitions": {
    "ErrorResponse": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "properties": {
          "type": "object",
          "properties": {
            "error": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Error message"
                }
              }
            }
          }
        },
        "example": {
          "type": "object",
          "properties": {
            "error": {
              "type": "string",
              "example": "Validation error"
            }
          }
        }
      }
    },
...
    "WalletBalance": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "properties": {
          "type": "object",
          "properties": {
            "balance": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "number"
                },
                "description": {
                  "type": "string",
                  "example": "Wallet balance"
                }
              }
            },
            "currency": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Currency code (e.g., NGN)"
                }
              }
            }
          }
        },
        "example": {
          "type": "object",
          "properties": {
            "balance": {
              "type": "number",
              "example": 12500.5
            },
            "currency": {
              "type": "string",
              "example": "NGN"
            }
          }
        }
      }
    },
    "WalletSendRequest": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "required": {
          "type": "array",
          "example": [
            "amount",
            "pin"
          ],
          "items": {
            "type": "string"
          }
        },
        "properties": {
          "type": "object",
          "properties": {
            "amount": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "number"
                },
                "description": {
                  "type": "string",
                  "example": "Transfer amount"
                }
              }
            },
            "pin": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Transaction PIN"
                }
              }
            },
            "channel": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Use `bank` for bank transfer."
                }
              }
            },
            "accountNumber": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Recipient account number (bank transfer)"
                }
              }
            },
            "bankCode": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Bank code (bank transfer)"
                }
              }
            },
            "accountName": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Recipient account name (bank transfer)"
                }
              }
            },
            "to": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Recipient email or phone for internal transfer."
                }
              }
            }
          }
        },
        "example": {
          "type": "object",
          "properties": {
            "amount": {
              "type": "number",
              "example": 5000
            },
            "pin": {
              "type": "string",
              "example": "1234"
            },
            "channel": {
              "type": "string",
              "example": "bank"
            },
            "accountNumber": {
              "type": "string",
              "example": "0123456789"
            },
            "bankCode": {
              "type": "string",
              "example": "058"
            },
            "accountName": {
              "type": "string",
              "example": "Ada Lovelace"
            }
          }
        }
      }
    },
    "WalletSendResponse": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "properties": {
          "type": "object",
          "properties": {
            "message": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Result message"
                }
              }
            },
            "reference": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Transfer reference"
                }
              }
            },
            "status": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Transfer status"
                },
                "enum": {
                  "type": "array",
                  "example": [
                    "pending",
                    "success"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "example": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "Transfer initiated"
            },
            "reference": {
              "type": "string",
              "example": "TX-abc123"
            },
            "status": {
              "type": "string",
              "example": "pending"
            }
          }
        }
      }
    },
    "WalletReceiveRequest": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "required": {
          "type": "array",
          "example": [
            "amount"
          ],
          "items": {
            "type": "string"
          }
        },
        "properties": {
          "type": "object",
          "properties": {
            "amount": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "number"
                },
                "description": {
                  "type": "string",
                  "example": "Requested amount"
                }
              }
            },
            "note": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Optional note for recipient"
                }
              }
            }
          }
        },
        "example": {
          "type": "object",
          "properties": {
            "amount": {
              "type": "number",
              "example": 1500
            },
            "note": {
              "type": "string",
              "example": "Lunch refund"
            }
          }
        }
      }
    },
    "WalletReceiveResponse": {
      "type": "object",
      "properties": {
        "type": {
          "type": "string",
          "example": "object"
        },
        "properties": {
          "type": "object",
          "properties": {
            "message": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Result message"
                }
              }
            },
            "reference": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "description": {
                  "type": "string",
                  "example": "Request reference"
                }
              }
            }
          }
        },
        "example": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "Money request created"
            },
            "reference": {
              "type": "string",
              "example": "REQ-xyz123"
            }
          }
        }
      }
    },
...
  }
}